pyproject.toml
[project]
name = "kgk-compe-5th"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "ipython>=9.2.0",
    "japanize-matplotlib>=1.1.3",
    "jupyterlab>=4.4.1",
    "lightgbm>=4.6.0",
    "mypy>=1.15.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "ruff>=0.11.7",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "taskipy>=1.14.1",
    "wandb>=0.19.10",
]

[tool.taskipy.tasks]
lint = "uv run ruff check src"
format = "uv run ruff format src"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "T201", # print
    "COM812", # missing-trailing-comma
]
unfixable = [
    "F401", # unused-import
    "F841", # unused-variable
]
pydocstyle.convention = "google"

[tool.ruff.lint.pylint]
max-args = 6
